Projects/View Funding: style Issues list
This does not use `Card/Issue`, despite the similar styling, for two
reasons:

* Some of the content of each are different – such as the dot-separated
  items at the bottom
* The styling differs slightly

I determined that reusing `Card/Issue` would require too much change to
that component for now.

If we want to align these components more later, we can now see the
differences more clearly as we decide how to approach that merger.