Projects: Simplify "SearchFilterAction" row
This removes the Viewport component, which added an `onResize` listener
to the page in order to alter the order of the items.

With CSS Grid, we don't need to actually reorder the markup in order to
change the placement of elements based on a media query. This
accomplishes the same behavior without using Viewport.

Other styling/markup issues address while here:

* Make the "Clear Filters" button an actual button
* Vertically center the active filters and Clear Filters button
* Run `prettier` on all edited files