Rewards: fix menu button styling
On narrow screens, the menu button in the Rewards app was located mostly
off-screen to the left. This puts it in a sensible location, with
margin-right that matches the Voting app from Aragon One.

This also greatly simplifies the styles so they're easier to see and
override, if specific apps need to override them. Hopefully other apps
don't need to override them, though!