Projects/ViewFunding: move logic to custom hook
BountyContextMenu doesn't want to know about the specifics of opening a
panel. It just wants to call a function.

This locates the logic for how to open a specific panel next to the
PanelManager component. This gives us a clean place to which to relocate
other panel-opening logic, rather than putting it all in App.js.