Projects/App.js: relocate newProject & newIssue
Perplexingly, without the change to `LoadingAnimation` to render an
inline svg, this resulted in cryptic Parcel errors like:

> Uncaught (in promise) Error: Cannot find module 'components/Shared/assets/svg/ethereum-loading.svg'
>    at newRequire (VM856 app.e31bb0bc.js:37)

Another way to fix this issue is to remove the dynamic imports from
`PanelManager`.

After hours of research, I'm not sure how the rest of this change
resulted in such an error, but I'm moving on for now.