Projects: improve useGithubAuth hook
App.js does not need `githubCurrentUser`, so there's no reason to fire
the `useEffect` hook and its background request to GitHub on initial
page load. This allows us to simplify the hook and only return
`githubCurrentUser`, instead of also returning `client`.

In addition, this also improves the custom hook itself:

* only run the effect if `token` is defined; return early otherwise
* since `setClient` fires asynchronously, `client` is not immediately
  defined; this caused 401 errors

Note that every component that needs this hook will re-fire the
`useEffect` function; they are not sharing global state for
`githubCurrentUser`. If the user opens many of the panels & pages that
need `githubCurrentUser`, this wastes their bandwidth. On slow
connections, this could become problematic. We may eventually want to
move this variable to a more global context, so only the first component
that needs `githubCurrentUser` fires the call to GitHub.