Projects/Issues: clarify & remove node0 checks
We were checking for the existence of `data && data.node0` in two
separate places.

By inverting the order of the render function – first check for
`loading` and `error`, and only if they're not the case, return the
actual content – it became clear why we needed this check. Unlike a
traditional pagination solution, this infinite-scroll requires us to
always render what we already have. "What we already have" is saved to
`this.state`. We only need to re-run `flattenIssues` if new data is
available from `Query`.