Dot Voting: clean end-of-vote watching
* Move logic to custom hook
* Use useEffect hook
* Only run useEffect hook when `votes` or `voteTime` change
  (`voteTime` should only change once, after api initialization, but if
  this happens after `votes` is initialized, we want to make sure we
  clear timers from when `voteTime` was `0` and set new timers)
* Use local variable `timeouts` to keep track of timeouts set within the
  context of the `useEffect` run
* Return `cleanup` closure function which has access to local
  `timeouts`; React will run this every time it run the effect to clear
  previous timers before setting new ones.

This new effect will run _every time `votes` changes_. If we are sloppy
with how frequently we update this array in our `script.js`, this will
be inefficient. However, I believe this is the safest way to write this
code. Should we have such innefficiencies, I believe we should try to
address them by updating script.js, not this code.

Note also that the `+ 5000` has gone away. This was present previously
to (incorrectly) deal with a bug caused by the use of `setInterval`. Now
that we use `setTimeout`, we do not need it.