Discussions: rm changes to Dot Voting
This is focusing this branch to add the minimal amount of code.

The changes in apps/dot-voting will cause significant conflicts when we
integrate Discussions with the Dot Voting design system upgrade. The
easiest path forward is to remove the changes entirely from this branch,
so they can be added only after the design system upgrade.