Discussions (#1050)
* paste in discussion app

* include discussions in npm start:dev build cmd

* discussion app working

* implement discussions as a useless forwarder

* temp forwarding through discussions

* fix state malfunctions from new aragon-api

* add sample external tx intent

* start integration with new forwarder API

* initial discussion state builder module framework

* build state for discussions in module

* initial integration into dot voting

* log discussion data

* associate discussion thread with dot vote

* display discussion panel and posts

* implement post functionality

* sort discussion posts before serving to react

* small fixes

* disable husky for time being

* programmatically load discussions

* update discussions as they occur

* bug fix

* Discussions: rm some UI that's not in design

* Discussions: extract form to CommentForm

* Clean up styles
* Improve behavior

* Discussions: DiscussionPost → Comment; styling

* Discussions/CommentForm: Add label

* Discussions/CommentForm: add hint text

* Discussions: add UI for edit & reply

These buttons look good, but don't do anything yet

* Discussions: Hook up frontend of comment editing

It needs a working save function in Discussion.js, but the rest is in
place and it behaves properly on the frontend.

* Discussions: allow cancelling comment add/edit

Also, rename `save` prop to `onSave`

* Discussions: confirm cancel

The `40px` bottom margin moved from CommentForm to Discussions because
we don't want the extra space to appear within an under-edit CommentCard

* Discussions: add placeholder Reply behavior

* Discussions/Comment: use actual currentUser

* Discussions/UI: silence propTypes errors

* update package.json with date-fns

* Discussions/Comment: add delete button

Not currently hooked up to API

* remove unused code

* Discussions: remove Reply

We are punting on this feature for now.

Now that Reply has been removed, these are moving to the right, and not
taking up any extra space below the content

* Discussions/Comment: confirm delete

After first click, show "Confirm delete" message. Only delete after
clicking that.

* add hide functionality

* Discussions: address some accessibility concerns

For some reason the `aria-live` elements here seem to get registered
twice, at least by Voice Over. This causes the updated content to be
read twice.

* Click "Delete", Voice Over speaks out "Confirm delete; Confirm delete"
* Click "Edit", Voice Over speaks out the new card contents, then at the
  end says "Cancel; Cancel"
* Click "Cancel", Voice Over speaks out "Confirm Cancel; Confirm Cancel"

I was unable to determine why this happens. The most relevant
information I found is a slightly different problem affecting nvda, a
different screen reader: https://github.com/nvaccess/nvda/issues/7996

* Discussions: render Markdown

* Move the Markdown comment used by Projects' IssueDetail to shared/ui
* Include react-markdown in package.json for Discussions
* use same Markdown component in both places
* fix Hint text to explain how to make things bold in Markdown

* update aragon api versions

* overwrite files that should not have been touched

* remove unused allocations files

* remove untouched project apps files

* one more untouched projects file

* remove untouched rewards files

* correct import to markdown file in projects

* reset dot voting untouched files

* more untouched code removal

* add a useHandshake hook

* fix dot voting lint errors

* use handshake in discussion api

* update packages for new api

* use autark ipfs node

* update radspec

* add discussion documentation

* discussions Readme updates

* remove discussions as a forwarder from dot voting

* Discussions: rm changes to Dot Voting

This is focusing this branch to add the minimal amount of code.

The changes in apps/dot-voting will cause significant conflicts when we
integrate Discussions with the Dot Voting design system upgrade. The
easiest path forward is to remove the changes entirely from this branch,
so they can be added only after the design system upgrade.

* Discussions: rm apps/rewards/package-lock.json

This probably got added by accident; it is not related to Discussions

* Discussions: revert unrelated package.json changes

These changes are unrelated to getting Discussions working; removing
them to make Discussions play more nicely with other branches.

* Discussions: rm obsolete Markdown component

This now lives in shared/ui/components/misc, and should only be used
from that location

* Discussions: rm apps/dot-voting/package-lock.json

This file should not be in this branch and was probably added by
accident.

* Discussions: fix linting errors and &/&& typos

Most of the linting errors were complaining about use of `now`. This [can
cause security problems][1], if using `now` for security-related
functions. We are not doing that here, so I am guessing that our use of
`now` is safe, and am therefore ignoring these lint warnings explicitly.

  [1]: https://consensys.github.io/smart-contract-best-practices/recommendations/#timestamp-dependence

* Fix lint and remove package-lock files

* Update lint script to use correct local config for each app

* hardcode 100 discussion thread Ids

* Discussions: add placeholder UI

This UI shows up when you select Discussions in the sidebar. We have
nothing to show for the Discussions app itself at this point in time,
and now we explain that here.

* Discussions: npm run lint:fix

* update contract