Dot Voting: minParticpationPct→globalMinQuorum
In the contract, the variable called `minParticpationPct` was renamed to
`globalMinQuorum`. This necessitates a change in some places
(`apps/dot-voting/app/utils/vote-settings.js`) and makes it
merely good practice in others.

Yes, this now means we have `vote.globalMinQuorum`, which looks exactly
as strange as it should. I like that it's now obvious that we're storing
this information in a weird place. We can now store that information
somewhere more expectable in a follow-up change.