local dev: stub useAragonApi for Dot Voting
Dot Votes are not created from within the Dot Voting app! To get around
this, I extended shared/api-react to add the custom functions to
`window.api`, and log a message to the console explaining that they're
there.

Also, there was a race condition that caused an invalid UI state, which
is fixed by always initializing appState from `db.fetchData`.