Allocations: fix creating new account/budget
We're in a halfway state right now. We don't have new contracts yet, but
we're starting to get new UI.

This keeps the new UI working with the current contracts.

The comparison of a raw JS Number object with MIN_AMOUNT, which is a
BigNumber, was also failing intermittently.

Finally, a step value of `1` is a bit large for ETH. It has been reduced
to `0.1`, which might still be too large.

This does NOT fix the issue that when you first press the up arrow,
while the Amount field is focused, the value to show up is the `min`,
`1e-18`! This value is a bit surprising. To fix it, we would need to
remove the `min` value from the DOM field, and enforce it only in JS.
That decision requires design discussion and is outside of the scope of
this small change.