Dot Voting: match slider values to actual vote
We weight the vote based on the user's `votingPower`, which is their
balance at the time the vote was created. We were already saving the
correct value, but we were not displaying it in all the sliders.

In a situation where I put 1% of my voting power to one option, and 0 to
the rest, it would say "You: 100%" in the badge, and when editing would
initialize the slider to 100% rather than 1%.

This fixes it.