Projects: fix initialState
`cachedState` now starts off, on the first run, as `null`. The attempt
to set the default param does not work, though, because the value is
intentionally set, not missing. So instead of defaulting within the
argument signature, we need to set the default at the first place where
we update our cache.