Projects/script: don't update state unnecessarily
This improves the performance of our app in general, but is crucial now
that we listen to all events from the StandardBounties contract. We want
to ignore all of these events, leaving state unchanged.

We could improve this even further, as shown in the `github` handlers.
We only need to mutate the specific branch of our state tree that
requires updating, and then React's algorithm that calculates whether it
needs to render only gets calculated for the components that make use of
that specific branch. However, this gets us past the bottleneck
introduced by full-state-rewrites-on-every-event-from-StandardBounties
bottleneck, and that's the main improvement we need right now.