PlanningKit.sol: update hard-coded StandardBounties address
This address changes depending on the content of our contracts and the
order in which they are deployed. This is accurate for running
`npm run publish:http` right now.

Until we get the new template in, which supports making this value a
variable, we will need to keep making PRs like this when new contracts
get merged.