Authenticator: Use postMessage origin to read URL
Since Authenticator is not allowed to read `opener.location.pathname`
directly, here's how this version works:

1. Projects opens a popup with the GitHub auth page
2. Our GitHub OAuth App is configured to redirect to the deployed
   Authenticator app (the index.html file here)
3. The Authenticator pings the `opener` (the Projects app)
4. Projects listens for this ping, and pings right back
5. Authenticator listens for pings from `opener`, and reads `origin`
   from these ping events.

This DOES NOT WORK for Aragon apps. It would work for apps that are not
loaded within an iframe, or for iframes that have [the
`allow-same-origin` property set in their `sandbox` attribute][1] (changing
this security setting is [not in the cards][2] for Aragon apps). But for
regular iframes such as any Aragon app, when they send `postMessage` to
the popup, the popup tries to read `message.origin`, and all it gets is
`"null"`.

IPFS hash for this version: QmaQFvQoGnWpsfLAumBJdSxV2zs9E6W5DPNB1fjg1QPxCu

  [1]: https://stackoverflow.com/questions/37838875/postmessage-from-a-sandboxed-iframe-to-the-main-window-origin-is-always-null
  [2]: https://github.com/aragon/aragon/blob/master/src/security/configuration.js#L22