Projects: stop making GitHub queries in script.js
Since we can't be confident that the user is logged in at the time
events are read from the blockchain,

And since we can't be sure we'll see these events again due to how
aragonAPI's store caching works,

We want everything in the appState reducer (aka script.js) to rely only
on blockchain and IPFS data.

This means that the other data in `repos` needs to be loaded in on the
frontend, once we know that the user has logged into GitHub.

Note that this architecture uses a `globalDecoratedRepos` variable,
defined at the top level of the new `hooks/useRepos` file. This ensures
that the first frontend component that needs to decorate repos (which
happens to be App.js) will update this file-global variable, and
subsequent components that need decorated repos can avoid unnecessary
API calls to GitHub.

Co-authored-by: PeterMPhillips <github@pmphillips.ca>