Allocations: Remove "Deactivate" option
Fixes #1474

This removes the "Deactivate" option from the Budget context menu, as
well as everything that's there to support it. This might be more than
we need, but if we don't have firm plans for when we will start
supporting "deactivate", then I think it's best to err toward "remove
more" rather than "remove less".