Left or right positioned DatePicker
In order to attached the DatePicker to the parent input without the DatePicker being cutoff because it overflows the parent element. You now have to pass the 'position' prop to the DateInput component. The default value if 'right' which is the behaviour it had before. If you pass 'left' it positions itself to the left edge of the input. Additionally, for this to work correctly, the parent element must have the css 'position: relative;' or 'position: absolute;'