Changed 'position' prop name and fixed widths of side-by-side date inputs
The 'position' prop name was changed to 'horizontal' and I added a 'vertical' prop so that you can attach the DatePicker at the top or bottom of the DatePicker component. Default is 'top'. Also made it so it attaches to its input rather than the input's container. Also there were some visual issues for the side-by-side inputs on mobile screens. On iPad they were spaced too far, and on small screens they were too large and overflowed. This was fixed as well.