chore: remove no-unused-modules rule
This rule causes all of our linting IDE integrations to move very
slowly, and I think commenting this out for now will enhance our
workflows. Note that I'm only commenting it out because I think it's
very useful when we want to optimize our builds. HOwever, for day-to-day
bugfixing and dev work it's probably not necessary