Refactored template
The commit changes the template to create a DAO across 3 transactions. It gives the option to add an additional token and token manager that can optionally control the Voting and Dot Voting apps. Additionally allows the creation of a membership token and the second token may be a reputation type with control of transfers handled by a WhitelistOracle.