AragonUI 1.0.0 upgrade
Upgraded all apps and followed (and improved) the test script on each
of them. The test script was run twice on each app, before and after
the upgrade, to make sure to detect only those bugs that were specific
to the upgrade. Only found one bug caused by the upgrade, which is
solved in this PR, and also found three more bugs that aren't related
to this upgrade and weren't reported before. Created tickets for
them.