Allocations/App.js: simplify setPanel interface
The most interesting part of this change is the removal of the
INITIAL_STATE resets in NewAllocation and NewBudget.

Note that the entire panel gets closed after these actions, making the
resets unnecessary. Closing the panel removes it entirely from the DOM,
so when the user creates a follow-up budget or allocation, the form will
be re-initialized from scratch.

Now that the `closePanel` behavior has a delay between closing the panel
and removing the panel from the DOM, these resets started causing
problems. The only reason they avoided errors before was because
we were trying to update unmounted components.