Open submission flow added
Requirement: to be able to submit work for bounties without
being an approved (via Application Review) owner.

On UI side it is done by Switch added to Fund Issue form.
The Switch sets openSubmission bool, which ends up in IPFS
data file, and it is later retrieved from there. Hence changes
to the store.

openSubmission variable is to be used in changes made in other
tickets, it doesn't do anything special in this one.