Removed bounty info from IssueDetails
Because all the info now lives in Bounty widget (former Status widget).
Also: cleaned up IssueDetails component - a lot of code was no longer
required. PanelComponents got touched, because Issue Title is a very
reusable component - in all panels, and issue details as well.
What needed to be done was restricting Link to only issue's ID
(it used to be whole div - including GH icon), and adding extra
children prop to the component, so I could do this:

<IssueTitleLink issue={issue}>
   {dot}
   <IconInfo
...

as per design. One note: we only download open issues right now.
What happens when an issue a bounty is attached to becomes closed...
That requires a separate ticket.