Open submission flow added (#1743)
* Open submission flow added

Requirement: to be able to submit work for bounties without being an approved (via Application Review) owner.

On UI side it is done by Switch added to Fund Issue form. The Switch sets openSubmission bool, which determines api function to call to create the bounty: addBounties vs addBountiesNoAssignment. The difference between them is that the latter set bounty's assignee field to 0xF{40}, and when that bounty is then downloaded from contract the store recognizes that address and sets issue's openSubmission appropriately.

* Added OpenSubmission specific display details

Original work (replacing Status with Bounty widget) was based
on full funding flow. This change extends Bounty widget,
so it displays appropriate texts for OS bounties. It is done
in this PR, because it belongs to 'Open submission flow'
ticket, and original work already got merged.