Allocations/BudgetDetail: support inactive budget
I really don't like the shenanigans required to get the Grid to behave
properly when it's essentially deactivated (that is, on narrow screens),
but this was the simplest way I found to achieve acceptable styling with
minimal effort.

Also note that the budget name using `Text size="great"` has some
hard-coded bottom margin that we can't get rid of.

Other things:

* don't show `New allocation` in context menu for inactive budgets
* don't list inactive budgets in NewAllocation dropdown