Incorrect status field in Application request
Each Application (request for work on a given issue) had a `status`
field containing text version of the application status: Unreviewed,
Accepted, Rejected. However, because the application was saved
on IPFS in its entirety, including initial Unreviewed status,
it was ignoring further (post-request-review) status changes.

It was not a problem, because we never relied on the status being
correct. We use `review.approved` as the ultimately (and correct)
source of truth.

The choice was between fixing the status field and removing it.
The reasons for choosing to remove it instead of fixing are:
1. There is no analogous field in Work Submission,
2. It's more complicated to fix it - correct way of fixing it
   would require not storing it on IPFS == actively removing
   it from data we need to store on IPFS,
3. It's not in use anyway. We have another and equally correct
   way of determining the application status, and it is the same
   way we use for Work Submissions. Standards FTW.