Allocations/NewAllocation: fix when no budget id
The main problem was that we were calculating how to render AmountField
even when we didn't actually want to render it, and rendering
AmountField requires a budget to be set in order to calculate the
Remaining Funds.

This cleans up the render function by moving these inline renders to
their own components instead.