Projects: mv TabData to route; use named tabs
This fixes a bug. Since TABS in usePathSegments was hard-coded at three
values but `tabs` in App.js was variable in length based on the presence
of appState.repos, setting `selectedTab` to a number could cause bugs
when the two conflicted. This sets `selectedTab` to a string instead,
which always works correctly and is more understandable.

This also removes `tabData`, instead directly using `requestPath` and
setting `repoId` in the URL, which is then consumed by Content/Issues.