refactor(projects): move hook out of Issue card
Since we currently render many (100+) Issue cards in this list, we want
to make sure each one renders as quickly as possible. One way we can do
this is to reduce the number of hooks this component calls.

This makes it so the `usePathHelpers` hook is processed once, rather
than 100+ times.