refactor(projects): use GitHub `search` query
This will allow us to support free-form queries when the user types in
the search field. Paging issues within a `repository` lookup does not.

Note that this method of searching does not allow sorting by creation
date: https://help.github.com/en/github/searching-for-information-on-github/sorting-search-results

After playing around with the available sorting options, I thought
"updated" seemed the only useful one, in the context of the Projects
app. I couldn't figure out what "author-date" was actually sorting on,
but it doesn't seem to be the creation date of the issue.