refactor(projects): filter&sort bountyIssues locally
bountyIssues only live locally. As more logic moves to the GitHub query,
we want to make sure we keep around logic that allows us to provide a
good experience when filtering for issues by bounty status.

This adds back local sorting, but it only gets used when filtering for
bounty status. It also cleans up filtering logic to make it work for the
new desired paradigm.